+++ Source/MVCF/Features/Feature.cs
@@ -2,9 +2,8 @@ using System.Collections.Generic;
 using System.Linq;
 using HarmonyLib;
 using MVCF.Comps;
-using MVCF.Features.PatchSets;
 using MVCF.ModCompat;
-using MVCF.ModCompat.PatchSets;
+using MVCF.PatchSets;
 using MVCF.Utilities;
 using MVCF.VerbComps;
 using Verse;
