+++ Source/MVCF/MVCF.csproj
@@ -86,16 +86,16 @@
     <Compile Include="Features\Feature_RangedAnimals.cs" />
     <Compile Include="Features\Feature_Reloading.cs" />
     <Compile Include="Features\Feature_VerbComps.cs" />
-    <Compile Include="Features\PatchSets\PatchSet_Base.cs" />
-    <Compile Include="Features\PatchSets\PatchSet_Brawlers.cs" />
-    <Compile Include="Features\PatchSets\PatchSet_Debug.cs" />
-    <Compile Include="Features\PatchSets\PatchSet_DualWield.cs" />
-    <Compile Include="Features\PatchSets\PatchSet_Equipment.cs" />
-    <Compile Include="Features\PatchSets\PatchSet_TargetFinder.cs" />
+    <Compile Include="PatchSets\PatchSet_Base.cs" />
+    <Compile Include="PatchSets\PatchSet_Brawlers.cs" />
+    <Compile Include="PatchSets\PatchSet_Debug.cs" />
+    <Compile Include="PatchSets\PatchSet_DualWield.cs" />
+    <Compile Include="PatchSets\PatchSet_Equipment.cs" />
+    <Compile Include="PatchSets\PatchSet_TargetFinder.cs" />
     <Compile Include="ModCompat\DualWieldCompat.cs" />
-    <Compile Include="Features\PatchSets\PatchSet_RunAndGun.cs" />
-    <Compile Include="Features\PatchSets\PatchSet_Hunting.cs" />
-    <Compile Include="Features\PatchSets\PatchSet_BatteLog.cs" />
+    <Compile Include="PatchSets\PatchSet_RunAndGun.cs" />
+    <Compile Include="PatchSets\PatchSet_Hunting.cs" />
+    <Compile Include="PatchSets\PatchSet_BatteLog.cs" />
     <Compile Include="JobGiver_ManhunterRanged.cs" />
     <Compile Include="ManagedVerb.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
