+++ Source/MVCF/PatchSets/PatchSet_TargetFinder.cs
@@ -5,7 +5,7 @@ using RimWorld;
 using Verse;
 using Verse.AI;
 
-namespace MVCF.Features.PatchSets;
+namespace MVCF.PatchSets;
 
 public class PatchSet_TargetFinder : PatchSet
 {
